fix: optimize download page architecture tab switching performance
Resolves issue #223 where architecture tab switching had noticeable delays
on deployed environments (staging.rockylinux.org and rockylinux.org) due to
server-side navigation and expensive data processing on every render.

**Key Changes:**

• **Client-side state management**: Replaced router.push() with local useState
  and startTransition for instant tab switching without server round-trips

• **Cached data processing**: Moved expensive download data transformation
  to utils/downloadDataProcessor.ts with module-level caching to prevent
  reprocessing on every render

• **Shallow URL updates**: Used window.history.pushState to maintain URL
  state for shareability without triggering server navigation

• **Improved TypeScript**: Added proper type definitions replacing 'any'
  with structured interfaces for better type safety

**Performance Impact:**
- Eliminates server round-trip delays on deployed Vercel environments
- Reduces data processing overhead through intelligent caching
- Maintains all existing functionality including browser back/forward navigation
- Preserves URL state for deep-linking and shareability

**Technical Details:**
- Uses React.startTransition for smooth, non-blocking UI updates
- Optimized useState initialization with function form
- Removed unused imports and variables
- All changes pass lint and type checking
